+++ ssl/sech.c
@@ -6,6 +6,8 @@
  * https://www.openssl.org/source/license.html
  */
 
+#include <stdlib.h>
+#include <openssl/bio.h>
 #include <openssl/ssl.h>
 #include <openssl/ech.h>
 #include "sech_local.h"
@@ -45,6 +47,10 @@ int SSL_CTX_sech_decode_sni(SSL_CTX *ctx)
 
 int SSL_CTX_sech_symmetric_key(SSL_CTX *ctx, char *key)
 {
+    fprintf(stderr, "SECH: key %s\n", key);
+    OSSL_TRACE_BEGIN(TLS) {
+        BIO_printf(trc_out, "SECH: test trace 1");
+    } OSSL_TRACE_END(TLS);
     if (key == NULL) {
         OSSL_TRACE_BEGIN(TLS) {
             BIO_printf(trc_out, "SECH: Supplied symmetric key is NULL\n");
@@ -64,6 +70,7 @@ int SSL_CTX_sech_symmetric_key(SSL_CTX *ctx, char *key)
         } OSSL_TRACE_END(TLS);
         return 0;
     }
+
     OSSL_TRACE_BEGIN(TLS) {
         BIO_printf(trc_out, "SECH: Symmetric key ascii len %i\n", asciilen);
     } OSSL_TRACE_END(TLS);
@@ -80,6 +87,13 @@ int SSL_CTX_sech_symmetric_key(SSL_CTX *ctx, char *key)
         char hex[3] = {key[2*i], key[2*i+1], '\0'};
         ctx->sech.symmetric_key[i] = (char) strtol(hex, NULL, 16);
     }
-    fprintf(stderr, "SECH: symmetric key: %s\n", ctx->sech.symmetric_key);
+    for(int i = 0; i < numBytes; ++i) {
+        fprintf(stderr, "%02X ",(unsigned char) ctx->sech.symmetric_key[i]);
+    }
+    fprintf(stderr,"\n");
+    OSSL_TRACE_BEGIN(TLS) {
+        fprintf(stderr, "SECH: test trace 2 fprintf");
+        BIO_printf(trc_out, "SECH: test trace 2");
+    } OSSL_TRACE_END(TLS);
     return 1;
 }
